+++ packages/zod-openapi/test/index.test.ts
@@ -33,31 +33,20 @@ describe('Basic - params', () => {
   const ParamsSchema = z.object({
     id: z
       .string()
-      .transform((val, ctx) => {
-        const parsed = parseInt(val)
-        if (isNaN(parsed)) {
-          ctx.addIssue({
-            code: z.ZodIssueCode.custom,
-            message: 'Not a number',
-          })
-          return z.NEVER
-        }
-        return parsed
-      })
+      .min(4)
       .openapi({
         param: {
           name: 'id',
           in: 'path',
         },
-        example: 123,
-        type: 'integer',
+        example: '12345',
       }),
   })
 
   const UserSchema = z
     .object({
-      id: z.number().openapi({
-        example: 123,
+      id: z.string().openapi({
+        example: '12345',
       }),
       name: z.string().openapi({
         example: 'John Doe',
@@ -67,6 +56,13 @@ describe('Basic - params', () => {
       }),
     })
     .openapi('User')
+
+  const HeadersSchema = z.object({
+    // Header keys must be in lowercase
+    authorization: z.string().openapi({
+      example: 'Bearer SECRET',
+    }),
+  })
 
   const ErrorSchema = z
     .object({
@@ -81,6 +77,7 @@ describe('Basic - params', () => {
     path: '/users/{id}',
     request: {
       params: ParamsSchema,
+      headers: HeadersSchema,
     },
     responses: {
       200: {
@@ -136,17 +133,21 @@ describe('Basic - params', () => {
   })
 
   it('Should return 200 response with correct contents', async () => {
-    const res = await app.request('/users/123')
+    const res = await app.request('/users/12345', {
+      headers: {
+        Authorization: 'Bearer TOKEN',
+      },
+    })
     expect(res.status).toBe(200)
     expect(await res.json()).toEqual({
-      id: 123,
+      id: '12345',
       age: 20,
       name: 'Ultra-man',
     })
   })
 
   it('Should return 400 response with correct contents', async () => {
-    const res = await app.request('/users/abc')
+    const res = await app.request('/users/123')
     expect(res.status).toBe(400)
     expect(await res.json()).toEqual({ ok: false })
   })
@@ -162,7 +163,7 @@ describe('Basic - params', () => {
           User: {
             type: 'object',
             properties: {
-              id: { type: 'number', example: 123 },
+              id: { type: 'string', example: '12345' },
               name: { type: 'string', example: 'John Doe' },
               age: { type: 'number', example: 42 },
             },
@@ -181,10 +182,16 @@ describe('Basic - params', () => {
           get: {
             parameters: [
               {
-                schema: { type: 'integer', example: 123 },
+                schema: { type: 'string', example: '12345', minLength: 4 },
                 required: true,
                 name: 'id',
                 in: 'path',
+              },
+              {
+                schema: { type: 'string', example: 'Bearer SECRET' },
+                required: true,
+                name: 'authorization',
+                in: 'header',
               },
             ],
             responses: {
@@ -1111,6 +1118,60 @@ describe('Path normalization', () => {
           },
         },
       })
+    })
+  })
+})
+
+describe('Context can be accessible in the doc route', () => {
+  const app = new OpenAPIHono<{ Bindings: { TITLE: string } }>()
+
+  app.openapi(
+    createRoute({
+      method: 'get',
+      path: '/no-content',
+      responses: {
+        204: {
+          description: 'No Content',
+        },
+      },
+    }),
+    (c) => {
+      return c.body(null, 204)
+    }
+  )
+
+  app.doc('/doc', context => ({
+    openapi: '3.0.0',
+    info: {
+      version: '1.0.0',
+      title: context.env.TITLE,
+    },
+  }))
+
+  it('Should return with the title set as specified in env', async () => {
+    const res = await app.request('/doc', null, { TITLE: 'My API' })
+    expect(res.status).toBe(200)
+    expect(await res.json()).toEqual({
+      openapi: '3.0.0',
+      info: {
+        version: '1.0.0',
+        title: 'My API',
+      },
+      components: {
+        schemas: {},
+        parameters: {},
+      },
+      paths: {
+        '/no-content': {
+          get: {
+            responses: {
+              204: {
+                description: 'No Content',
+              },
+            },
+          },
+        },
+      },
     })
   })
 })